Adding documentation for default min and max replicass
Fix for https://github.com/ray-project/ray/issues/50222. 

There were long attempts to verify this locally. However building documentation locally failed continuously as mentioned in this defect. As suggested in ray-contributors to make  PR, trying the same. 
The changes are one line adding documentation just to default() method. The intent is to see how alternatively the changes can be verified. 

Signed-off-by: Jagan Molleti <hi2mjb@gmail.com>